split common helper into 2 + new directory
The current `test/test_helper/common.bash` is quite good, but rather long.
I therefore split the setup process into a separate file. I dislike that
the directoty `common.bash` is in is called `test_helper` - the `test_`
prefix is redundant. Therefore, I came up with an interim solution that
provides the new approach to the updated tests while the "old" tests can
use the "old" `common.bash`. The new helper files are located under
`test/helper/` which drops the `test_` prefix.

The new and updated helper use the new naming convention for ENV
variables, such as `CONTAINER_NAME` or `IMAGE_NAME`. I encountered a
situation before in which the updated tests would fail because there
were collisions of ENV names in the tests, for example with `CONTAINER_NAME`.